Make load() no longer bypasses the config
When accessing name.xml, it ignores the configuration file.
It should be a mistake and in some cases it will pause the server, when the server is not able to access the site xml.yuttyann44581.net.
It should be controlled by the configuration so the plugin can still work well in those cases.
Thanks. I'm not very understand in your PR rules and the modifies not been tested, so it just a suggection of code.